+++ schnorr_fun/tests/against_c_lib.rs
@@ -69,7 +69,7 @@ proptest! {
         let sig = secp.sign_schnorr_no_aux_rand(&msg, &keypair);
         let schnorr = Schnorr::<Sha256,Bip340NoAux>::default();
         let fun_keypair = schnorr.new_keypair(key);
-        let fun_msg = Message::<Public>::raw(&msg);
+        let fun_msg = Message::raw(&msg);
         let fun_sig: secp256k1::schnorr::Signature = schnorr.sign(&fun_keypair, fun_msg).into();
         prop_assert_eq!(fun_sig, sig, "they produce the same signatures");
     }
@@ -82,7 +82,7 @@ proptest! {
         let fun_pk = secp256k1::XOnlyPublicKey::from_keypair(&keypair).0.into();
         let sig = secp.sign_schnorr_with_aux_rand(&msg, &keypair, &aux_rand);
         let schnorr = Schnorr::<Sha256,_>::verify_only();
-        let fun_msg = Message::<Public>::raw(&msg);
+        let fun_msg = Message::raw(&msg);
         prop_assert!(schnorr.verify(&fun_pk, fun_msg, &sig.into()));
     }
 }